Adding sleeper job, renaming web_driver class
